user profile lookup added. user recommendation algorithm updated.

+++ src/main/java/com/erkutoguz/moviever_backend/dto/response/RecommendedMovieResponse.java
@@ -1,4 +1,4 @@
 package com.erkutoguz.moviever_backend.dto.response;
 
-public record RecommendedMovieResponse(int movieId, double score) {
+public record RecommendedMovieResponse(long movieId, double score) {
 }
